refactor(web): decompose DocsPage into 5 colocated components (#453)
DocsPage went from 905 lines down to 303 (−67%, just over the issue's 300-line target). Five new components under 'components/docs/':

- DocsMermaid (244L) — MermaidBlock + MermaidLightbox + SandboxedSvg + per-theme palettes. The biggest extraction by far. Mermaid is initialised at module-import time with the user's current theme; MermaidBlock re-initialises on theme change before each render.
- DocsMarkdownComponents (139L) — markdownComponents object spread into <ReactMarkdown components={...}>: CopyButton-wrapped PreBlock, mermaid-aware CodeBlock, slug-injecting H1-H4. Exports the shared slugify() because the TOC needs IDs that match the heading IDs.
- DocsReleaseAccordion (139L) — ReleaseAccordion + VersionBadge (release notes per-version expander + the inline version eyebrow).
- DocsSidebar (112L) — collapsible left rail. Auto-expands the section containing the active doc, re-collapses on user click.
- DocsTableOfContents (69L) — sticky right minimap. Skips h1, indents h2-h4 by depth.

Behavior is identical. DocsPage now only carries: the page's own state (active doc, scroll-spy active heading, frontmatter parse, copy-doc handler) and the layout shell (sidebar + content + TOC + RELEASES/VERSION_BADGE markdown injection points).

Per the issue's 'one PR per page' guidance: SkillDetailPage went out in #651 (1133→868); this PR handles DocsPage; PlaygroundPage (813L) still needs its own PR.
